nfs41: layoutget may get layout stateid instead of open stateid
Motivation:
When client hols a layout and requests a new one on the same file, then
layout stateid is used instead of open stateid. This situation may happen
when client originally got or requested layout for fagment of a file or
replacing RO layout with RW one.

Modification:
If stateid is derived from opens stateid, get the original stateid and
use it when talking to device manager.

Result:
accepting layout stateid when processing layoutget.

Acked-by: 
Target: master